manager/requirements.txt
flask
redis
pygit2
requests
gitpython
